138-PV/Radar-Superadmin-adds-generic-Radar-group
On successful addition go back to list of Radar units. Otherwise the
form is still filled in and someone will press "Add" button again
thinking that it hasn't worked.
Needs validation on the form: all fields are mandatory, email must be
well formed.
If you add a new radar group which has a duplicate unitcode of an
existing group then a 404 error appears. Need to check for this and give
a warning on the page so the user can correct it.